requirements.txt
fastapi~=0.111.1
torch==2.4
torchvision==0.19
opencv-python~=4.10.0.84
numpy~=1.26.4
pillow>=10
keras>=2.4
requests~=2.32.3
pydantic~=2.9.2
uvicorn~=0.30.6
torchsampler~=0.1.2